sample-api/src/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1

opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0

prometheus-client==0.20.0
httpx==0.27.0